Update BT: MaximumDepthofBinaryTree.java

+++ BT: MaximumDepthofBinaryTree.java
@@ -5,6 +5,6 @@ public class MaximumDepthofBinaryTree {
         int leftDepth = maxDepth(root.left);
         int rightDepth = maxDepth(root.right);
         
-        return (int)Math.max(leftDepth, rightDepth) + 1;
+        return Math.max(leftDepth, rightDepth) + 1;
     }
 }
